+++ src/Controller/Webhook.php
@@ -2,37 +2,28 @@
 
 namespace Ginger\EmsPay\Controller;
 
-use Symfony\Component\Routing\Annotation\Route;
 use Shopware\Core\Framework\Routing\Annotation\RouteScope;
+use Symfony\Component\Routing\Annotation\Route;
+use Shopware\Storefront\Controller\StorefrontController;
+use Symfony\Component\HttpFoundation\Response;
+use Ginger\ApiClient;
+use Ginger\EmsPay\Service\ClientBuilder;
 use Shopware\Core\Checkout\Order\Aggregate\OrderTransaction\OrderTransactionStateHandler;
 use Shopware\Core\Framework\Context;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
-use Shopware\Core\System\SystemConfig\SystemConfigService;
-use Ginger\EmsPay\Service\Helper;
+
 /**
  * @RouteScope(scopes={"storefront"})
  */
 
-class Webhook extends AbstractController
+class Webhook extends StorefrontController
 {
     /**
-     * @var Helper
-     */
-
-    private $helper;
-
-    /**
-     * @var \Ginger\ApiClient
+     * @var ApiClient
      */
 
     private $ginger;
 
-    /**
-     * @var SystemConfigService
-     */
-
-    private $EmsPayConfig;
 
     /**
      * @var OrderTransactionStateHandler
@@ -43,20 +34,17 @@ class Webhook extends AbstractController
     /**
      * Webhook constructor.
      * @param OrderTransactionStateHandler $transactionStateHandler
-     * @param SystemConfigService $systemConfigService
-     * @param Helper $helper
+     * @param ClientBuilder $clientBuilder
      */
 
-    public function __construct(OrderTransactionStateHandler $transactionStateHandler, SystemConfigService $systemConfigService, Helper $helper)
+    public function __construct(OrderTransactionStateHandler $transactionStateHandler, ClientBuilder $clientBuilder)
     {
         $this->transactionStateHandler = $transactionStateHandler;
-        $this->EmsPayConfig = $systemConfigService->get('EmsPay.config');
-        $this->helper = $helper;
-        $this->ginger = $this->helper->getGignerClinet($this->EmsPayConfig['emsOnlineApikey'], $this->EmsPayConfig['emsOnlineBundleCacert']);
+        $this->ginger = $clientBuilder->getClient();
     }
 
     /**
-     * @Route("/EmsPay/Webhook", name="frontend.checkout.emspay.webhook", options={"seo"="false"}, methods={"GET"})
+     * @Route("/EmsPay/Webhook", methods={"GET"})
      */
 
     public function webhook(Request $request, Context $context)
@@ -69,15 +57,19 @@ class Webhook extends AbstractController
 
             $ginger_order = $this->ginger->getOrder($request_content->order_id);
             $shopware_order_id = $ginger_order['extra']['sw_order_id'];
-
-            if ($ginger_order['status'] != 'new')
-            $this->transactionStateHandler->reopen($shopware_order_id, $context);
+            try{
+                $this->transactionStateHandler->reopen($shopware_order_id, $context);
+            } catch(\Exception $exception) {
+                print('Already opened');
+            }
 
         switch ($ginger_order['status']) {
+                case 'accepted':
                 case 'completed' : $this->transactionStateHandler->paid($shopware_order_id, $context); break;
                 case 'cancelled' : $this->transactionStateHandler->cancel($shopware_order_id, $context); break;
                 case 'error' : $this->transactionStateHandler->fail($shopware_order_id, $context); break;
                 case 'processing' : $this->transactionStateHandler->process($shopware_order_id, $context); break;
             }
+        return new Response('Webhook successful processed',200,[]);
     }
 }