+++ src/Service/Gateway.php
@@ -3,16 +3,17 @@
 namespace Ginger\EmsPay\Service;
 
 use Ginger\ApiClient;
+use Ginger\EmsPay\Exception\EmsPluginException;
 use Shopware\Core\Checkout\Order\Aggregate\OrderTransaction\OrderTransactionStateHandler;
 use Shopware\Core\Checkout\Payment\Cart\AsyncPaymentTransactionStruct;
 use Shopware\Core\Checkout\Payment\Cart\PaymentHandler\AsynchronousPaymentHandlerInterface;
 use Shopware\Core\Checkout\Payment\Exception\AsyncPaymentProcessException;
 use Shopware\Core\Checkout\Payment\Exception\CustomerCanceledAsyncPaymentException;
+use Shopware\Core\Framework\DataAbstractionLayer\EntityRepositoryInterface;
 use Shopware\Core\Framework\Validation\DataBag\RequestDataBag;
 use Shopware\Core\System\SalesChannel\SalesChannelContext;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
-use Shopware\Core\System\SystemConfig\SystemConfigService;
 
 class Gateway implements AsynchronousPaymentHandlerInterface
 {
@@ -35,18 +36,45 @@ class Gateway implements AsynchronousPaymentHandlerInterface
     private $transactionStateHandler;
 
     /**
+     * @var EntityRepositoryInterface
+     */
+
+    private $orderRepository;
+
+    /**
+     * @var mixed
+     */
+
+    private $use_webhook;
+
+    /**
+     * @var
+     */
+
+    private $clientBuilder;
+
+
+    /**
      * Gateway constructor.
+     * @param EntityRepositoryInterface $orderRepository
      * @param OrderTransactionStateHandler $transactionStateHandler
-     * @param SystemConfigService $systemConfigService
+     * @param ClientBuilder $clientBuilder
      * @param Helper $helper
      */
 
-    public function __construct(OrderTransactionStateHandler $transactionStateHandler, SystemConfigService $systemConfigService, Helper $helper)
+    public function __construct
+    (
+        EntityRepositoryInterface $orderRepository,
+        OrderTransactionStateHandler $transactionStateHandler,
+        ClientBuilder $clientBuilder,
+        Helper $helper
+    )
     {
+        $this->orderRepository = $orderRepository;
         $this->transactionStateHandler = $transactionStateHandler;
         $this->helper = $helper;
-        $EmsPayConfig = $systemConfigService->get('EmsPay.config');
-        $this->ginger = $this->helper->getGignerClinet($EmsPayConfig['emsOnlineApikey'], $EmsPayConfig['emsOnlineBundleCacert']);
+        $this->clientBuilder = $clientBuilder;
+        $this->use_webhook = $this->clientBuilder->getConfig()['emsOnlineUseWebhook'];
     }
 
     /**
@@ -57,17 +85,21 @@ class Gateway implements AsynchronousPaymentHandlerInterface
         RequestDataBag $dataBag,
         SalesChannelContext $salesChannelContext
     ): RedirectResponse {
+
         // Method that sends the return URL to the external gateway and gets a redirect URL back
         try {
+            $this->ginger = $this->clientBuilder->getClient($transaction->getOrderTransaction()->getPaymentMethod()->getDescription());
             $pre_order = $this->processOrder($transaction,$salesChannelContext);
             $order = $this->ginger->createOrder($pre_order);
+
+            if($order['status'] == 'error') {
+                $this->transactionStateHandler->fail($transaction->getOrderTransaction()->getId(), $salesChannelContext->getContext());
+                throw new EmsPluginException(current($order['transactions'])['reason']);
+            }
         } catch (\Exception $e) {
-            print_r($e->getMessage()); exit;
-            throw new AsyncPaymentProcessException(
-                $transaction->getOrderTransaction()->getId(),
-                'An error occurred during the creating the EMS Online order' . PHP_EOL . $e->getMessage()
-            );
+            throw new EmsPluginException($e->getMessage());
         }
+
         // Redirect to external gateway
         return new RedirectResponse(
             isset($order['order_url']) ? $order['order_url'] : current($order['transactions'])['payment_url']
@@ -82,14 +114,19 @@ class Gateway implements AsynchronousPaymentHandlerInterface
         Request $request,
         SalesChannelContext $salesChannelContext
     ): void {
-        $transactionId = $transaction->getOrderTransaction()->getId();
+        $this->ginger = $this->clientBuilder->getClient($salesChannelContext->getPaymentMethod()->getDescription());
         $order = $this->ginger->getOrder($_GET['order_id']);
         $context = $salesChannelContext->getContext();
-        $transaction->getOrderTransaction()->setCustomFields(['ginger_order_id' => $order['id']]);
         $paymentState = $order['status'];
         if (!($this->helper::SHOPWARE_STATES_TO_GINGER[$transaction->getOrderTransaction()->getStateMachineState()->getTechnicalName()] == $paymentState))
             switch ($paymentState) {
-            case 'completed' : $this->transactionStateHandler->paid($transaction->getOrderTransaction()->getId(), $context); break;
+            case 'completed' :
+                $this->helper->saveGingerOrderId(
+                    $transaction->getOrderTransaction()->getId(),
+                    $order['id'],
+                    $this->orderRepository,
+                    $context);
+                $this->transactionStateHandler->paid($transaction->getOrderTransaction()->getId(), $context); break;
             case 'cancelled' : $this->transactionStateHandler->cancel($transaction->getOrderTransaction()->getId(), $context); break;
             case 'new' : $this->transactionStateHandler->reopen($transaction->getOrderTransaction()->getId(), $context); break;
             case 'processing' : $this->transactionStateHandler->process($transaction->getOrderTransaction()->getId(), $context); break;
@@ -97,26 +134,24 @@ class Gateway implements AsynchronousPaymentHandlerInterface
                $message ='Error during transaction';
                $message .= isset(current($order['transactions'])['reason']) ? ':'.current($order['transactions'])['reason'].'.' : '.';
                $message .= '<br> Please contact support.';
-            print_r($message);exit;
-            throw new CustomerCanceledAsyncPaymentException(
-                $transactionId,
-                (current($order['transactions'])['reason'])
-            ); break;
+
+            throw new EmsPluginException($message);
         }
     }
 
     private function processOrder($transaction,$sales_channel_context): array
     {
+        $issuer_id = $sales_channel_context->getPaymentMethod()->getCustomFields()['issuer_id'] ?? null;
         return array_filter([
             'amount' => $this->helper->getAmountInCents($transaction->getOrder()->getAmountTotal()),                                                     // Amount in cents
             'currency' => $sales_channel_context->getCurrency()->getIsoCode(),                                                                           // Currency
             'merchant_order_id' => $transaction->getOrder()->getOrderNumber(),                                                                           // Merchant Order Id
             'description' => $this->helper->getOrderDescription($transaction->getOrder()->getOrderNumber(),$sales_channel_context->getSalesChannel()),   // Description
             'customer' => $this->helper->getCustomer($sales_channel_context->getCustomer()),                                                             // Customer information
-            'order_lines' => $this->helper->getOrderLines($sales_channel_context,$transaction->getOrder()),                          // Order Lines
-            'transactions' => $this->helper->getTransactions($sales_channel_context->getPaymentMethod(),$this->ginger->getIdealIssuers()),               // Transactions Array
+            'order_lines' => $this->helper->getOrderLines($sales_channel_context,$transaction->getOrder()),                                              // Order Lines
+            'transactions' => $this->helper->getTransactions($sales_channel_context->getPaymentMethod(), $issuer_id),                                    // Transactions Array
             'return_url' => $transaction->getReturnUrl(),                                                                                                // Return URL
-            'webhook_url' => $this->helper->getWebhookUrl(),                                                                                             // Webhook URL
+            'webhook_url' => $this->use_webhook ? $this->helper->getWebhookUrl() : null,                                                                                             // Webhook URL
             'extra' => $this->helper->getExtraArray($transaction->getOrderTransaction()->getId()),                                                       // Extra information
             'payment_info' => [],                                                                                                                        // Payment info
         ]);