+++ src/Service/Helper.php
@@ -2,12 +2,13 @@
 
 namespace Ginger\EmsPay\Service;
 
-use Ginger\ApiClient;
-use Ginger\Ginger;
+use Shopware\Core\Framework\Log\LoggerFactory;
+use Monolog\Processor\WebProcessor;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsFilter;
 
 class Helper
 {
-
     const SHOPWARE_STATES_TO_GINGER =
         [
             'paid' => 'completed',
@@ -50,44 +51,16 @@ class Helper
     const DEFAULT_CURRENCY = 'EUR';
 
     /**
-     *  Default Ginger endpoint
-     */
-
-    const GINGER_ENDPOINT = 'https://api.online.emspay.eu';
-
-    /**
      * Constructor of the class which includes ginger-php autoload
      */
 
-    public function __construct(){
-        include (dirname(__FILE__)."/../Vendor/vendor/autoload.php");
-    }
-
     /**
-     * create a gigner clinet instance
-     *
-     * @param string $apiKey
-     * @param boolean $useBundle
-     * @return ApiClient
+     *  Logger Factory
      */
-    public function getGignerClinet($apiKey, $useBundle = false)
-    {
-        return Ginger::createClient(
-            self::GINGER_ENDPOINT,
-            $apiKey,
-            $useBundle ?
-                [
-                    CURLOPT_CAINFO => self::getCaCertPath()
-                ] : []
-        );
-    }
+    private $loggerFactory;
 
-    /**
-     *  function get Cacert.pem path
-     */
-
-    protected static function getCaCertPath(){
-        return dirname(__FILE__).'/../Vendor/assets/cacert.pem';
+    public function __construct(LoggerFactory $loggerFactory){
+        $this->loggerFactory = $loggerFactory;
     }
 
     /**
@@ -212,6 +185,15 @@ class Helper
     }
 
     /**
+     * @param $payment
+     * @return mixed
+     */
+
+    private function translatePaymentMethod($payment){
+        return !is_null($payment) ? self::SHOPWARE_TO_EMS_PAYMENTS[explode('emspay_',$payment)[1]] : null;
+    }
+
+    /**
      * Get transactions array which includes required API information
      *
      * @param $payment
@@ -221,7 +203,7 @@ class Helper
 
     public function getTransactions($payment,$issuer){
 
-        $ginger_payment = self::SHOPWARE_TO_EMS_PAYMENTS[explode('emspay_',$payment->getDescription())[1]];
+        $ginger_payment = $this->translatePaymentMethod($payment->getDescription());
 
         return array_filter([
             array_filter([
@@ -317,7 +299,7 @@ class Helper
      */
 
     public function getOrderLines($sales,$order){
-        if (!in_array($sales->getPaymentMethod(),['emspay_klarnapaylater','emspay_afterpay']))
+        if (!in_array($sales->getPaymentMethod()->getDescription(),['emspay_klarnapaylater','emspay_afterpay']))
         {
             return null;
         }
@@ -327,5 +309,51 @@ class Helper
         }
         $order->getShippingCosts()->getUnitPrice() > 0 ? array_push($order_lines,self::getShippingLines($sales,$order)) : null;
         return $order_lines;
+    }
+
+    /**
+     * Function saveEMSLog
+     * Writes a log to a file /app/var/log/ems_plugin_%environment%-%current date%. If there are more than 7 logging files in the log directory, removes the oldest
+     *
+     * @param $msg
+     * @param $context
+     */
+    public function saveEMSLog($msg, $context) {
+        $ems_logger = $this->loggerFactory->createRotating('ems_plugin', 7);
+        $ems_logger->pushProcessor(new WebProcessor());
+        $ems_logger->error($msg, $context);
         }
+
+    /**
+     * Save the Ginger order id into Shopware Order for keep link between Ginger order ID
+     *
+     * @param $payment_method
+     * @param $orderTransactionId
+     * @param $ems_order_id
+     * @param $orderRepository
+     * @param $context
+     * @return mixed
+     */
+
+    public function saveGingerOrderId($orderTransactionId,$ems_order_id,$orderRepository,$context){
+        //Search the Shopware Order using transaction id.
+        $orderCriteria = new Criteria();
+        $orderCriteria->addFilter(new EqualsFilter('transactions.id', $orderTransactionId));
+        $order = $orderRepository->search($orderCriteria, $context)->first();
+
+        //Update customFields.
+        $order_custom_fields = $order->getCustomFields();
+        $order_custom_fields = array_merge(
+            empty($order_custom_fields) ? [] : $order_custom_fields,
+            ['ems_order_id' => $ems_order_id]
+        );
+
+        //Return updated Shopware order.
+        return $orderRepository->update(
+            [
+                ['id' => $order->getId(), 'customFields' => $order_custom_fields],
+            ],
+            $context
+        );
+    }
 }