+++ CHANGELOG.md
@@ -3,3 +3,19 @@
 ** 1.0.0 **
 
 * Initial Version
+
+** 1.1.0 **
+
+* Has been implemented: 
+    * orderCapture functionality.
+    * Webhook settings option.
+    * Test API-Key options in settings for 
+        * Klarna Pay Later.
+        * Afterpay.
+    * Test IP Address options in settings for 
+        * Afterpay.
+        * Klarna Pay Later.
+    * Country Validation.
+    * Error Logging & Error Processing.
+    * Issuer Id Selector field.
+    * Afterpay birthday select field.